+++ video_xblock/backends/wistia.py
@@ -12,6 +12,7 @@ import requests
 import babelfish
 
 from video_xblock import BaseVideoPlayer
+from video_xblock.constants import TranscriptSource
 from video_xblock.utils import ugettext as _
 from video_xblock.exceptions import VideoXBlockException
 
@@ -64,7 +65,14 @@ class WistiaPlayer(BaseVideoPlayer):
 
         Brightcove videos require Brightcove Account id.
         """
-        fields_list = super(WistiaPlayer, self).advanced_fields
+        return super(WistiaPlayer, self).advanced_fields
+
+    @property
+    def three_pm_fields(self):
+        """
+        Tuple of VideoXBlock fields to display on `3PlayMedia transcripts` panel.
+        """
+        fields_list = super(WistiaPlayer, self).three_pm_fields
         # Add `token` field before `threeplaymedia_file_id`
         fields_list.insert(fields_list.index('threeplaymedia_file_id'), 'token')
         return fields_list
@@ -206,6 +214,7 @@ class WistiaPlayer(BaseVideoPlayer):
                         'label': lang_label,
                         'url': '',
                         'text': text,
+                        'source': TranscriptSource.DEFAULT,
                     })
             # If captions do not exist for a video, the response will be an empty JSON array.
             # Reference: https://wistia.com/doc/data-api#captions_index