+++ video_xblock/video_xblock.py
@@ -5,29 +5,31 @@ All you need to provide is video url, this XBlock does the rest for you.
 """
 
 import datetime
-import json
 import httplib
+import json
 import logging
 import os.path
-import requests
 
+import requests
+from webob import Response
 from xblock.core import XBlock
 from xblock.fields import Scope, Boolean, String, Dict
 from xblock.fragment import Fragment
 from xblock.validation import ValidationMessage
 from xblockutils.studio_editable import StudioEditableXBlockMixin
 
-from webob import Response
-
-from .backends.base import BaseVideoPlayer
-from .constants import PlayerName
-from .exceptions import ApiClientError
-from .mixins import ContentStoreMixin, LocationMixin, PlaybackStateMixin, SettingsMixin, TranscriptsMixin
-from .workbench.mixin import WorkbenchMixin
-from .settings import ALL_LANGUAGES
-from .fields import RelativeTime
-from .utils import render_template, render_resource, resource_string, ugettext as _
 from . import __version__
+from .backends.base import BaseVideoPlayer
+from .constants import PlayerName, TranscriptSource
+from .exceptions import ApiClientError
+from .fields import RelativeTime
+from .mixins import ContentStoreMixin, LocationMixin, PlaybackStateMixin, SettingsMixin, TranscriptsMixin
+from .settings import ALL_LANGUAGES
+from .utils import (
+    create_reference_name, filter_transcripts_by_source, normalize_transcripts,
+    render_resource, render_template, resource_string, ugettext as _,
+)
+from .workbench.mixin import WorkbenchMixin
 
 log = logging.getLogger(__name__)
 
@@ -149,24 +151,9 @@ class VideoXBlock(
         help=_(
             'Default transcripts are uploaded automatically from a video platform '
             'to the list of available transcripts.<br/>'
-            '<b>Note: "Video API Token" should be given in order to make auto fetching possible.</b>'
+            '<b>Note: valid "Video API Token" should be given in order to make auto fetching possible.</b><br/>'
+            'Advice: disable transcripts displaying on your video service to avoid transcripts overlapping.'
         ),
-        resettable_editor=False
-    )
-
-    threeplaymedia_apikey = String(
-        default='default',
-        display_name=_('API Key'),
-        help=_('You can generate a client token following official documentation of your video platform\'s API.'),
-        scope=Scope.content,
-        resettable_editor=False
-    )
-
-    threeplaymedia_file_id = String(
-        default='default',
-        display_name=_('File Id'),
-        help=_('3playmedia file id for download bind transcripts.'),
-        scope=Scope.content,
         resettable_editor=False
     )
 
@@ -315,7 +302,7 @@ class VideoXBlock(
                 display_name=self.display_name,
                 usage_id=self.usage_id,
                 handout=self.handout,
-                transcripts=self.route_transcripts(self.transcripts),
+                transcripts=self.route_transcripts(),
                 download_transcript_allowed=self.download_transcript_allowed,
                 download_video_url=self.get_download_video_url(),
                 handout_file_name=self.get_file_name_from_path(self.handout),
@@ -332,8 +319,8 @@ class VideoXBlock(
         """
         Private method to fetch/update default transcripts.
         """
+        log.debug("Default transcripts updating...")
         # Prepare parameters necessary to make requests to API.
-        log.debug("Updating default transcripts...")
         video_id = player.media_id(self.href)
         kwargs = {'video_id': video_id}
         for k in self.metadata:
@@ -343,11 +330,13 @@ class VideoXBlock(
             self.account_id is not self.fields['account_id'].default  # pylint: disable=unsubscriptable-object
         if is_not_default_account_id:
             kwargs['account_id'] = self.account_id
+
         # Fetch captions list (available/default transcripts list) from video platform API
         try:
             default_transcripts, transcripts_autoupload_message = player.get_default_transcripts(**kwargs)
         except ApiClientError:
             default_transcripts, transcripts_autoupload_message = [], _('Failed to fetch default transcripts.')
+        log.debug("Autofetch message: '{}'".format(transcripts_autoupload_message))
         # Default transcripts should contain transcripts of distinct languages only
         distinct_default_transcripts = player.clean_default_transcripts(default_transcripts)
         # Needed for frontend
@@ -368,7 +357,7 @@ class VideoXBlock(
         player = self.get_player()
         languages = [{'label': label, 'code': lang} for lang, label in ALL_LANGUAGES]
         languages.sort(key=lambda l: l['label'])
-        transcripts = json.loads(self.transcripts) if self.transcripts else []
+        transcripts = normalize_transcripts(json.loads(self.transcripts)) if self.transcripts else []
         download_transcript_handler_url = self.runtime.handler_url(self, 'download_transcript')
         auth_error_message = ''
 
@@ -377,7 +366,7 @@ class VideoXBlock(
         # whilst for Wistia, a sample authorised request is to be made to ensure authentication succeeded,
         # since it is needed for the auth status message generation and the player's state update with auth status.
         if self.token:
-            _auth_data, auth_error_message = self.authenticate_video_api(self.token)
+            _auth_data, auth_error_message = self.authenticate_video_api(self.token.encode(encoding='utf-8'))
 
         initial_default_transcripts, transcripts_autoupload_message = self._update_default_transcripts(
             player, transcripts
@@ -386,26 +375,33 @@ class VideoXBlock(
         # Prepare basic_fields and advanced_fields for them to be rendered
         basic_fields = self.prepare_studio_editor_fields(player.basic_fields)
         advanced_fields = self.prepare_studio_editor_fields(player.advanced_fields)
-        log.debug("Fetched default transcripts: {}".format(self.default_transcripts))
+        log.debug("Fetched default transcripts: {}".format(initial_default_transcripts))
         context = {
+            'advanced_fields': advanced_fields,
+            'auth_error_message': auth_error_message,
+            'basic_fields': basic_fields,
             'courseKey': self.course_key,
             'languages': languages,
-            'transcripts': transcripts,
-            'download_transcript_handler_url': download_transcript_handler_url,
-            'default_transcripts': self.default_transcripts,
-            'initial_default_transcripts': initial_default_transcripts,
-            'auth_error_message': auth_error_message,
-            'transcripts_autoupload_message': transcripts_autoupload_message,
-            'basic_fields': basic_fields,
-            'advanced_fields': advanced_fields,
             'player_name': self.player_name,  # for players identification
             'players': PlayerName,
+            'sources': TranscriptSource.to_dict().items(),
+            # transcripts context:
+            'transcripts': transcripts,
+            'transcripts_fields': self.prepare_studio_editor_fields(player.trans_fields),
+            'three_pm_fields': self.prepare_studio_editor_fields(player.three_pm_fields),
+            'transcripts_type': '3PM' if self.threeplaymedia_streaming else 'manual',
+            'default_transcripts': self.default_transcripts,
+            'enabled_default_transcripts': filter_transcripts_by_source(transcripts),
+            'initial_default_transcripts': initial_default_transcripts,
+            'transcripts_autoupload_message': transcripts_autoupload_message,
+            'download_transcript_handler_url': download_transcript_handler_url,
         }
 
         fragment.content = render_template('studio-edit.html', **context)
         fragment.add_css(resource_string("static/css/student-view.css"))
         fragment.add_css(resource_string("static/css/transcripts-upload.css"))
         fragment.add_css(resource_string("static/css/studio-edit.css"))
+        fragment.add_css(resource_string("static/css/studio-edit-accordion.css"))
         fragment.add_javascript(resource_string("static/js/runtime-handlers.js"))
         fragment.add_javascript(resource_string("static/js/studio-edit/utils.js"))
         fragment.add_javascript(resource_string("static/js/studio-edit/studio-edit.js"))
@@ -429,7 +425,7 @@ class VideoXBlock(
         save_state_url = self.runtime.handler_url(self, 'save_player_state')
         transcripts = render_resource(
             'static/html/transcripts.html',
-            transcripts=self.route_transcripts(self.transcripts)
+            transcripts=self.route_transcripts()
         ).strip()
         return player.get_player_html(
             url=self.href, account_id=self.account_id, player_id=self.player_id,
@@ -553,7 +549,7 @@ class VideoXBlock(
                 'has_list_values': False,
                 'type': 'string',
             }
-        elif field_name in ('handout', 'transcripts', 'default_transcripts', 'token', 'threeplaymedia_apikey'):
+        elif field_name in ('handout', 'transcripts', 'default_transcripts', 'token'):
             info = self.initialize_studio_field_info(field_name, field, field_type=field_name)
         else:
             info = self.initialize_studio_field_info(field_name, field)
@@ -742,14 +738,15 @@ class VideoXBlock(
             response (dict): Data on a default transcript, fetched from a video platform.
 
         """
+        log.debug("Uploading default transcript with data: {}".format(data))
         player = self.get_player()
         video_id = player.media_id(self.href)
         lang_code = str(data.get(u'lang'))
         lang_label = str(data.get(u'label'))
+        source = str(data.get(u'source', ''))
         sub_url = str(data.get(u'url'))
-        log.debug("Current media ID: " + video_id)
-        # File name format is <language label>_captions_video_<video_id>, e.g. "English_captions_video_456g68"
-        reference_name = "{}_captions_video_{}".format(lang_label, video_id).encode('utf8')
+
+        reference_name = create_reference_name(lang_label, video_id, source)
 
         # Fetch default transcript
         unicode_subs_text = player.download_default_transcript(
@@ -770,6 +767,19 @@ class VideoXBlock(
             'success_message': success_message,
             'lang': lang_code,
             'url': external_url,
-            'label': lang_label
+            'label': lang_label,
+            'source': source,
         }
+        log.debug("Uploaded default transcript: {}".format(response))
         return response
+
+    def get_enabled_transcripts(self):
+        """
+        Get transcripts from different sources depending on current usage mode.
+        """
+        if self.threeplaymedia_streaming:
+            transcripts = list(self.fetch_available_3pm_transcripts())
+        else:
+            transcripts = json.loads(self.transcripts) if self.transcripts else []
+
+        return transcripts