+++ CHANGELOG.md
@@ -7,9 +7,37 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.9.0] - 2017-07-11
+
+### Added
+
+- Direct transcripts streaming from 3PlayMedia service.
+  Now Video Xblock will always use up-to-date transcripts from 3PM.
+
+### Changed
+
+- UI: Transcripts settings section now split into two panels.
+  To make more apparent how different options are related and which
+  transcripts are going to be diplayed.
+  - "Manual & default transcripts"
+  - "3PlayMedia transcripts"
+
+## [0.8.0] - 2017-06-30
+
+### Added
+
+- Vimeo: default transcripts fetching.
+
+## Changed
+
+- Use new 3PlayMedia API to fetch transcripts translations.
+- Increase test coverage. Which means better stability.
+
 ### Fixed
 
 - Bug preventing setting Video Platform API key. E.g. `BC_TOKEN` for Brightcove.
+- Transcripts collision bug, which prevented teachers to replace
+  transcript for a given language.
 
 ## [0.7.1] - 2017-05-18
 
@@ -197,4 +225,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [0.6.6]: https://github.com/raccoongang/xblock-video/compare/v0.6.5...v0.6.6
 [0.7.0]: https://github.com/raccoongang/xblock-video/compare/v0.6.6...v0.7.0
 [0.7.1]: https://github.com/raccoongang/xblock-video/compare/v0.7.0...v0.7.1
-[Unreleased]: https://github.com/raccoongang/xblock-video/compare/v0.7.1...HEAD
+[0.8.0]: https://github.com/raccoongang/xblock-video/compare/v0.7.1...v0.8.0
+[0.9.0]: https://github.com/raccoongang/xblock-video/compare/v0.8.0...v0.9.0
+[Unreleased]: https://github.com/raccoongang/xblock-video/compare/v0.9.0...HEAD